fix: 件数が増えたときにうまく取得できない問題を修正
