+++ src/atoms/tourActions.ts
@@ -14,24 +14,27 @@ export const tourState = atom<TourState>({
 })
 
 type TourActions = {
-  useReloadTour: () => () => void
-  useAddTour: () => (tour: Tour) => void
-  useUpdateTour: () => (tour: Tour) => void
+  useReloadTour: () => (callback?: () => void) => void
+  useAddTour: () => (tour: Tour, callback?: () => void) => void
+  useUpdateTour: () => (tour: Tour, callback?: () => void) => void
+  useDeleteTour: () => (tourId: string, callback?: () => void) => void
 }
 
 export const tourActions: TourActions = {
   useReloadTour: () =>
-    useRecoilCallback(({ set }) => () => {
-      chromeStorageActions.getAll<Tour>('tours', (tours) =>
+    useRecoilCallback(({ set }) => (callback?: () => void) => {
+      chromeStorageActions.getAll<Tour>('tours', (tours) => {
         set(tourState, () => {
           return {
             tours: tours,
           }
         })
-      )
+
+        callback()
+      })
     }),
   useAddTour: () =>
-    useRecoilCallback(({ set }) => (tour: Tour) => {
+    useRecoilCallback(({ set }) => (tour: Tour, callback?: () => void) => {
       chromeStorageActions.add<Tour>('tours', tour, () => {
         // TODO: reload と同じ処理
         chromeStorageActions.getAll<Tour>('tours', (tours) =>
@@ -41,10 +44,12 @@ export const tourActions: TourActions = {
             }
           })
         )
+
+        callback()
       })
     }),
   useUpdateTour: () =>
-    useRecoilCallback(({ set }) => (tour: Tour) => {
+    useRecoilCallback(({ set }) => (tour: Tour, callback?: () => void) => {
       chromeStorageActions.update<Tour>('tours', tour.id, tour, () => {
         // TODO: reload と同じ処理
         chromeStorageActions.getAll<Tour>('tours', (tours) =>
@@ -54,6 +59,23 @@ export const tourActions: TourActions = {
             }
           })
         )
+
+        callback()
+      })
+    }),
+  useDeleteTour: () =>
+    useRecoilCallback(({ set }) => (tourId: string, callback?: () => void) => {
+      chromeStorageActions.remove<Tour>('tours', tourId, () => {
+        // TODO: reload と同じ処理
+        chromeStorageActions.getAll<Tour>('tours', (tours) =>
+          set(tourState, () => {
+            return {
+              tours: tours,
+            }
+          })
+        )
+
+        callback()
       })
     }),
 }
